refactor(C3): extract PendingOpView → runtime/pending_op_view.py (#1792)
FP-0044 C-series stage 3. Extracts the `PendingOpView` frozen dataclass
verbatim from `reyn.runtime.session` (the god-file) into its own
`reyn.runtime.pending_op_view` module so the cross-channel pending-op view
lives apart from the Session god-class.

- Pure move, byte-identical class body (no behavior change, no rename).
- `pending_op_view.py` docstring is current-state only (what the module IS);
  this refactor narrative lives here in the commit, per the C-series docstring
  discipline (C1 #1790 R1).
- `UserIntervention` (the from_intervention arg type) imported under
  TYPE_CHECKING only — never at runtime, so zero cycle risk; the
  from_intervention body is duck-typed.
- Clean break, no re-export shim: the one real consumer
  (tests/test_pending_intervention_268.py) repointed
  `from reyn.runtime.session import PendingOpView`
  → `from reyn.runtime.pending_op_view import PendingOpView` (mixed line split
  so `Session` stays sourced from session). All TUI / slash references are
  duck-typed comments, not imports — untouched.
- session.py imports the type back for its own list_stalled_interventions /
  claim return types.
- Repo-wide 4-ref-class straggler grep (dotted-import / dotted-literal /
  segment-path / docstring+comment) = 0 references to the old location.
