+++ tests/test_pending_intervention_268.py
@@ -42,8 +42,9 @@ from datetime import datetime
 
 import pytest
 
+from reyn.runtime.pending_op_view import PendingOpView
 from reyn.runtime.services.intervention_registry import InterventionRegistry
-from reyn.runtime.session import PendingOpView, Session
+from reyn.runtime.session import Session
 from reyn.user_intervention import (
     InterventionAnswer,
     UserIntervention,
